Added new combinators many2 and many3
many2 is similar to many1 but requires at least 2 instances of the given
syntax to succed. It is used to diferentiate between blocks with a single
element and blocks with 2 or more elements.
many3 takes two parameters and succedes if it gets an arbitrary number of
instances defined in the first parameter, followed by a single instance of
the syntax in the second parameter, followed again by an arbitrary number
of instances of the first parameter. It also succedes if it recieves no
instance of the second parameter.
It was meant to be used for single element blocks, where curly braces are
optional, but it is not clear if the method can be implemented correctly.